Fixed width / height bug and finished render function.
A bug was crossing references to width and height and would crash if they were different numbers.

Render function shows the grid with '#' for 1 and a space for 0. Clear lambda added for refreshing the terminal, but is commented out of the render function for now since the game doesn't actually need to run continuously yet.